🐛 :: resolve strange seatId!!

+++ app/src/main/java/com/g3c1/oasis_android/feature_seat/data/dto/SeatDTO.kt
@@ -1,7 +1,7 @@
 package com.g3c1.oasis_android.feature_seat.data.dto
 
 data class SeatDTO(
-    val seatId: Int,
+    val seatId: Long,
     val seatNumber: Int,
     val severalPeople: Int,
     val x: Float,
